clean up: move AlwaysFailStorage to separate file
